Refactoring. Changed int to unsigned into in lots of places.

+++ Neuro.Tests/src/RandomTests.cpp
@@ -12,12 +12,12 @@ namespace NeuroTests
         {
             Random rng(101);
             vector<int> v1(30);
-            for (size_t i = 0; i < v1.size(); ++i)
+            for (uint i = 0; i < v1.size(); ++i)
                 v1[i] = rng.Next(10000);
 
             Random rng2(101);
             vector<int> v2(v1.size());
-            for (size_t i = 0; i < v2.size(); ++i)
+            for (uint i = 0; i < v2.size(); ++i)
                 v2[i] = rng2.Next(10000);
 
             Assert::IsTrue(v1 == v2);
@@ -27,12 +27,12 @@ namespace NeuroTests
         {
             Random rng(101);
             vector<float> v1(30);
-            for (size_t i = 0; i < v1.size(); ++i)
+            for (uint i = 0; i < v1.size(); ++i)
                 v1[i] = rng.NextFloat();
 
             Random rng2(101);
             vector<float> v2(v1.size());
-            for (size_t i = 0; i < v2.size(); ++i)
+            for (uint i = 0; i < v2.size(); ++i)
                 v2[i] = rng2.NextFloat();
 
             Assert::IsTrue(v1 == v2);
